+++ ledger/ledger-api-integration-tests/src/test/lib/scala/com/digitalasset/platform/apitesting/LedgerContext.scala
@@ -18,6 +18,8 @@ import com.digitalasset.ledger.api.v1.active_contracts_service.ActiveContractsSe
 import com.digitalasset.ledger.api.v1.active_contracts_service.ActiveContractsServiceGrpc.ActiveContractsService
 import com.digitalasset.ledger.api.v1.admin.party_management_service.PartyManagementServiceGrpc
 import com.digitalasset.ledger.api.v1.admin.party_management_service.PartyManagementServiceGrpc.PartyManagementService
+import com.digitalasset.ledger.api.v1.admin.package_management_service.PackageManagementServiceGrpc
+import com.digitalasset.ledger.api.v1.admin.package_management_service.PackageManagementServiceGrpc.PackageManagementService
 import com.digitalasset.ledger.api.v1.command_completion_service.CommandCompletionServiceGrpc
 import com.digitalasset.ledger.api.v1.command_completion_service.CommandCompletionServiceGrpc.CommandCompletionService
 import com.digitalasset.ledger.api.v1.command_service.CommandServiceGrpc
@@ -56,7 +58,6 @@ import scala.concurrent.{ExecutionContext, Future}
 import scala.language.implicitConversions
 import scala.util.Success
 import scala.concurrent.duration._
-
 import scalaz.syntax.tag._
 
 trait LedgerContext {
@@ -91,6 +92,7 @@ trait LedgerContext {
   def acsClient: ActiveContractSetClient
   def reflectionService: ServerReflectionGrpc.ServerReflectionStub
   def partyManagementService: PartyManagementService
+  def packageManagementService: PackageManagementService
 
   /**
     * resetService is protected on purpose, to disallow moving an instance of LedgerContext into an invalid state,
@@ -226,6 +228,9 @@ object LedgerContext {
 
     override def partyManagementService: PartyManagementService =
       PartyManagementServiceGrpc.stub(channel)
+
+    override def packageManagementService: PackageManagementService =
+      PackageManagementServiceGrpc.stub(channel)
   }
 
   object SingleChannelContext {