+++ docs/source/support/release-notes.rst
@@ -9,6 +9,35 @@ This page contains release notes for the SDK.
 HEAD — ongoing
 --------------
 
+DAML Assistant
+~~~~~~~~~~~~~~
+
+- Added `--install-assistant` flag to `daml install` command, changing the default
+  behavior of `daml install` to be "install the assistant whenever we are installing
+  a newer version of the SDK". Deprecated the `--activate` flag.
+
+DAML Studio
+~~~~~~~~~~~
+
+- Opening an already open scenario will now focus it rather than opening
+  it in a new empty tab which is never updated with results.
+
+.. _release-0-13-5:
+
+0.13.5 - 2019-06-19
+-------------------
+
+Release Procedure
+~~~~~~~~~~~~~~~~~
+
+- Fixes to the CI/CD release procedure.
+  See `#1755 <https://github.com/digital-asset/daml/issues/1755>__.`
+
+Sandbox
+~~~~~~~
+
+- Introduced a new API for package management.
+  See `#1311 <https://github.com/digital-asset/daml/issues/1311>`__.
 
 .. _release-0-13-4:
 
@@ -258,7 +287,6 @@ Navigator
   permitting large packages; e.g. pass ``--ledger-api-inbound-message-size-max 62914560``
   to extractor to get a 60MiB limit.
   See `#1520 <https://github.com/digital-asset/daml/pull/1520>`__.
-
 
 Sandbox
 ~~~~~~~
@@ -815,7 +843,6 @@ No user-facing changes.
        // Accessing field 'bar' of field 'foo' of the argument
        DamlLfValue.evalPath(argument, ["foo", "bar"])
        DamlLfValue.toJSON(argument).foo.bar
-
 
 .. _release-0-11-32: